pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "map_my_world"
version = "1.0.0"
description = "REST API para generar ubicaciones de lugares con interés y enlazarlas con categorías, con revisiones periódicas y endpoint sandbox con mapa interactivo"
authors = [
    {name = "Ernesto Vivas", email = "vivas.ernesto@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["fastapi", "rest-api", "maps", "locations", "categories", "reviews", "geolocation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "jinja2",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "black",
    "isort",
    "flake8",
]

[project.urls]
Homepage = "https://github.com/Ernesiunp1/map_my_world"
Repository = "https://github.com/Ernesiunp1/map_my_world"
Issues = "https://github.com/Ernesiunp1/map_my_world/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["map_my_world*"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
